requirements.txt
aiogram~=3.19.0
python-dotenv~=1.1.0
APScheduler~=3.11.0
pytz~=2025.2
httpx~=0.28.1
certifi==2024.2.2